Add intersection testing to JS cut 1 layout
This represents the first part of the JS translation to be
unambiguously better than the PostScript version: now it can tell you
where the hlines run into one or another object in the scene, which is
what I need to actually be able to cut those supports.  It’s done in a
straightforward, unoptimized brute-force nested-loops way.

I suppose the next step is to plot the intersections on the graph and
then try to come up with some kind of laser-cut shape for them, then
move on to laying out the laser-cut shapes in a plane for laser
cutting.